+++ src/test/java/eu/verdelhan/tailtest/indicator/tracker/WMATest.java
@@ -3,6 +3,7 @@ package eu.verdelhan.tailtest.indicator.tracker;
 import eu.verdelhan.tailtest.Indicator;
 import eu.verdelhan.tailtest.indicator.simple.ClosePrice;
 import eu.verdelhan.tailtest.sample.SampleTimeSeries;
+import java.math.BigDecimal;
 import static org.junit.Assert.assertEquals;
 import org.junit.Test;
 
@@ -11,7 +12,7 @@ public class WMATest {
 	public void testWMACalculate()
 	{
 		SampleTimeSeries series = new SampleTimeSeries(new double[] {1d, 2d, 3d, 4d, 5d, 6d});
-		Indicator<Double> close = new ClosePrice(series);
+		Indicator<BigDecimal> close = new ClosePrice(series);
 		Indicator<Double> wmaIndicator = new WMA(close, 3);
 		
 		assertEquals(1d, (double) wmaIndicator.getValue(0));
@@ -26,7 +27,7 @@ public class WMATest {
 	public void testWMACalculateJumpingIndex()
 	{
 		SampleTimeSeries series = new SampleTimeSeries(new double[] {1d, 2d, 3d, 4d, 5d, 6d});
-		Indicator<Double> close = new ClosePrice(series);
+		Indicator<BigDecimal> close = new ClosePrice(series);
 		Indicator<Double> wmaIndicator = new WMA(close, 3);
 		
 		assertEquals(32d/6, (double) wmaIndicator.getValue(5));
@@ -36,7 +37,7 @@ public class WMATest {
 	public void testWMACalculateWithTimeFrameGreaterThanSeriesSize()
 	{
 		SampleTimeSeries series = new SampleTimeSeries(new double[] {1d, 2d, 3d, 4d, 5d, 6d});
-		Indicator<Double> close = new ClosePrice(series);
+		Indicator<BigDecimal> close = new ClosePrice(series);
 		Indicator<Double> wmaIndicator = new WMA(close, 55);
 		
 		assertEquals(1d, (double) wmaIndicator.getValue(0));
